refactor(rollout): remove obsolete async engine layer
