+++ unirl/rollout/engine/composed/config.py
@@ -17,7 +17,7 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import Any, Optional
 
-from unirl.rollout.engine.synchronous import BaseEngineConfig
+from unirl.rollout.engine.base import BaseEngineConfig
 
 
 @dataclass
