+++ unirl/rollout/engine/sglang_diffusion/config.py
@@ -21,8 +21,8 @@ from typing import Any, Dict, Optional, Tuple
 from omegaconf import SI
 
 from unirl.config.require import require
+from unirl.rollout.engine.base import BaseEngineConfig
 from unirl.rollout.engine.ports import ReservedPorts
-from unirl.rollout.engine.synchronous import BaseEngineConfig
 
 
 @dataclass(frozen=True)
