+++ unirl/rollout/engine/sglang_diffusion/engine.py
@@ -4,7 +4,7 @@ A thin core over the backend seam: it names no concrete model (the adapter, pick
 from the registry by ``config.model_family``, owns the ``Sample`` → ``Sample``
 conversion) and no concrete backend (the seam owns the runtime). Weight sync is a
 :class:`WeightSync` component constructed over the seam; the offload lifecycle (a
-single flag) lives directly on the engine. The frozen ``synchronous.py`` surface is
+single flag) lives directly on the engine. The frozen ``base.py`` surface is
 implemented as thin forwards here — they must be real class attributes anyway
 (``Worker.call`` dispatches by name; ``@distributed`` binds the most-derived
 attribute) — which also absorbs the surface quirks (``track_prefix``) so the
@@ -12,7 +12,7 @@ component keeps clean signatures.
 
 One-shot construction: after ``__init__`` returns, the generator is spawned and the
 engine is usable. ``generate`` / ``sleep`` / ``wake_up`` re-apply ``@distributed``
-(the decorator is not inherited — see ``synchronous.py``).
+(the decorator is not inherited — see ``base.py``).
 """
 
 from __future__ import annotations
@@ -25,6 +25,7 @@ import torch
 
 from unirl.config.require import require
 from unirl.distributed.group.dispatch import Dispatch, distributed
+from unirl.rollout.engine.base import SyncRolloutEngine
 from unirl.rollout.engine.sglang_diffusion.adapters import get_adapter
 from unirl.rollout.engine.sglang_diffusion.backends import SGLangBackend
 from unirl.rollout.engine.sglang_diffusion.config import (
@@ -32,7 +33,6 @@ from unirl.rollout.engine.sglang_diffusion.config import (
     SGLangDiffusionPorts,
 )
 from unirl.rollout.engine.sglang_diffusion.weight_sync import WeightSync
-from unirl.rollout.engine.synchronous import SyncRolloutEngine
 from unirl.sde.noise import generate_latents
 from unirl.sde.runtime import ensure_sample_sigmas
 from unirl.types.noise_recipe import NoiseRecipe
