+++ unirl/rollout/engine/trainside/__init__.py
@@ -1,7 +1,7 @@
 """In-process rollout engine adapter for direct-sampling mode.
 
 Exposes a materialized ``models`` ``Pipeline`` as a
-:class:`unirl.rollout.engine.synchronous.BaseRolloutEngine`. Used when the
+:class:`unirl.rollout.engine.base.BaseRolloutEngine`. Used when the
 training Policy itself is the sampler (direct sampling, on-policy RL) —
 the rollout runs in the same Ray actor / Python process / GPU as
 training, so no worker subprocess and no weight sync are needed.
