+++ unirl/rollout/engine/trainside/config.py
@@ -13,7 +13,7 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-from unirl.rollout.engine.synchronous import BaseEngineConfig
+from unirl.rollout.engine.base import BaseEngineConfig
 
 
 @dataclass
