+++ unirl/rollout/engine/trainside/engine.py
@@ -18,7 +18,7 @@ from unirl.distributed.group.dispatch import Dispatch, distributed
 from unirl.models.types.ar import ARStage
 from unirl.models.types.diffusion import DiffusionStage
 from unirl.models.types.pipeline import Pipeline
-from unirl.rollout.engine.synchronous import SyncRolloutEngine
+from unirl.rollout.engine.base import SyncRolloutEngine
 from unirl.sde.runtime import FlowMatchSchedulePolicy, ensure_sample_sigmas
 from unirl.types.sample import Part, Sample
